NEW/requirements.txt
# Requirements for Browser Reader Dashboard
# Same as main project, but explicitly listed here

flask>=2.0.0
flask-socketio>=5.0.0
python-dotenv>=0.19.0
playwright>=1.40.0
aiohttp>=3.9.0

# OCR Reader requirements (optional - only if using OCR method)
pytesseract>=0.3.10
pillow>=10.0.0
mss>=9.0.0

# Install playwright browsers:
# playwright install chromium

# Install Tesseract OCR separately:
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Mac: brew install tesseract
# Linux: sudo apt-get install tesseract-ocr
